Automatic channel disconnection after socket write timeout
Sometimes a thread is blocked at writing to or flushing a socket. It
happens for example when the connection is stale. Usually it takes less
then 1 milisecond. Introduced improvement disconnect the channel
if writing an ISO message bytes to a socket and/or flushing it takes
longer then 2000 miliseconds. In most cases this value is enough.
It can be changed by 'socket-write-timeout' parameter.